+++ test/globals-shim.js
@@ -1,3 +1,4 @@
+import document from 'global/document';
 import window from 'global/window';
 import sinon from 'sinon';
 
@@ -6,5 +7,5 @@ window.sinon = sinon;
 
 // This may not be needed anymore, but double check before removing
 window.fixture = document.createElement('div');
-fixture.id = 'qunit-fixture';
-document.body.appendChild(fixture);
+window.fixture.id = 'qunit-fixture';
+document.body.appendChild(window.fixture);